python-jr/pyproject.toml
[tool.poetry]
name = "python-jr"
version = "2.0.1"
description = "python plugin system for jr, the data random generator"
authors = ["JR Team"]
readme = "README.md"
packages = [{include = "python_jr"}]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.66.0"
grpcio-health-checking = "^1.59.0"
grpcio-tools = "^1.66.0"
protobuf = "^5.27.3"
setuptools = "^73.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
